Fixed  palette reducer and RGBA color conversion
Improved SimilarDistanceReducer not approximating when
the palettes can be added directly.
Fixed SimilarDistance because the distance is the value that
palette1 differs to palette2 and it is usually different of distance
from palette2 to palette1.

Temp fixed in color conversion for RGBA color since in Emgu.CV
there is a bug where the property Red returns the Blue color and vice versa.