Fixed problem creating lineal sprite images.
An horizontal image can not be converted into
lineal because the image is divided in blocks of
different size (Obj) so, each block must be converted
separately into lineal. For this reason I have implemented
that in ColorQuantization directly.

Another way (I have just think about -.-') could be convert
into lineal or horizontal just after the output of the quantization.